fix(ios): resolve zone picker chips invisible on Applications page
The large navigation title (.navigationBarTitleDisplayMode(.large))
hijacked the first scrollable content for its collapse animation,
corrupting the zone picker's rendering. Replaced separate stacked
ScrollViews with a single List as the sole scroll container — zone
picker and filter chips are now List rows. Also replaced
Button+buttonStyle(.plain) with Text+onTapGesture to avoid further
rendering issues.

Deleted ZonePickerView.swift — both ApplicationListView and MapView
now inline their zone chips directly.

Also fixed 5 pre-existing SwiftLint opening_brace violations in
unrelated files that were blocking the pre-commit hook.
