+++ mobile/ios/packages/town-crier-data/Sources/Auth/Auth0AuthenticationService.swift
@@ -20,8 +20,7 @@ public struct Auth0Config: Sendable {
 }
 
 public final class Auth0AuthenticationService: TownCrierDomain.AuthenticationService,
-  @unchecked Sendable
-{
+  @unchecked Sendable {
   private let config: Auth0Config
   private let credentialsManager: CredentialsManager
