+++ mobile/ios/packages/town-crier-domain/Sources/Entities/EntitlementMap.swift
@@ -41,8 +41,8 @@ public enum EntitlementMap {
   }
 
   /// Returns whether the given tier grants the specified entitlement.
-  public static func hasEntitlement(_ entitlement: Entitlement, for tier: SubscriptionTier) -> Bool
-  {
+  public static func hasEntitlement(_ entitlement: Entitlement, for tier: SubscriptionTier)
+    -> Bool {
     entitlements(for: tier).contains(entitlement)
   }
